requirements.txt
mediapipe
opencv-python
numpy
pillow
tensorflow
